backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.7.8
langchain==0.0.348
langchain-openai==0.0.2
langchain-anthropic==0.0.1
pandas==2.1.4
numpy==1.25.2
pytest==7.4.3
pytest-asyncio==0.21.1